requirements.txt
Flask
flask-pymongo
flask-mail
werkzeug
pyjwt